update to migration content

+++ database/migrations/2023_05_22_045355_create_categories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->string("name");
             
-            $table->integer("category_id")->unsigned();
+            $table->bigInteger("category_id")->unsigned();
             $table->foreign("category_id")->references("id")->on("categories");
 
             $table->timestamps();
