+++ database/migrations/2023_05_23_045130_create_staff_logs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('staff_logs', function (Blueprint $table) {
             $table->id();
 
-            $table->integer("user_id")->unsigned();
+            $table->bigInteger("user_id")->unsigned();
             $table->foreign("user_id")->references("id")->on("users");
 
             $table->dateTime("created_at");
